feat(FRONT): separate the frindrequist to components/utils
