+++ Client/src/packages/feat-Account/components/index.ts
@@ -3,9 +3,21 @@ import UserCard from "./UserCard";
 import BoardCard from "./BoardCard";
 import MatchsHistoryCard from "./MatchsHistoryCard";
 import AchievemetsCard from "./AchievemetsCard";
+import SendFriendRequist from "./utils/frindrequist";
+import SendGameRequist from "./utils/gameRequist";
+import ListButton from "./utils/listButton";
 export * from "./utils/utils";
 export * from "./ProfileSlice";
 export * from "./thunks";
 export * from "./statsType";
 
-export { UserCard, BoardCard, MatchsHistoryCard, AchievemetsCard, Loading };
+export {
+  UserCard,
+  BoardCard,
+  MatchsHistoryCard,
+  AchievemetsCard,
+  Loading,
+  SendFriendRequist,
+  SendGameRequist,
+  ListButton,
+};
